2FA/requirements.txt
tensorflow-macos==2.10.0
keras==2.10.0
keras-vggface==0.6
keras-applications==1.0.8
numpy<2
scikit-learn
opencv-python
pillow
